api/requirements.txt
Flask
Flask-Cors
PyMySQL
PyJWT
